+++ ui/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @platforma-open/milaboratories.mixcr-clonotyping.ui
 
+## 1.20.0
+
+### Minor Changes
+
+- 070fc0b: Min quality checkbox, refactored error correction, dep updates
+
+### Patch Changes
+
+- Updated dependencies [070fc0b]
+  - @platforma-open/milaboratories.mixcr-clonotyping-2.model@1.21.0
+
 ## 1.19.0
 
 ### Minor Changes